updated docker container workflow to trigger on pull_request for main/dev and push for database branch.
each branch will get two tags. one for the branch name and "latest". the latest tag will follow whichever container is literally the latest. 
/TP-3T/dockerimage:database &
/TP-3t/dockerimage:latest
will both point to the same image if database is pushed to.
if you then pull_request into dev,
/TP-3T/dockerimage:dev &
/TP-3T/dockerimage:latest 
will both point to the same image.
latest will follow what is actually the latest. the branch tag will point to the latest image for that branch.